Refac: seperate function
