+++ pythainlp/util/__init__.py
@@ -102,8 +102,8 @@ from pythainlp.util.normalize import (
     remove_tonemark,
     remove_zw,
     reorder_vowels,
-    remove_repeat_consonants
 )
+from pythainlp.util.removerepeatconsonants import remove_repeat_consonants
 from pythainlp.util.numtoword import bahttext, num_to_thaiword
 from pythainlp.util.strftime import thai_strftime
 from pythainlp.util.thai import (
